os_name {sessioninfo}	R Documentation

Human readable name of the current operating system

Description

For example Windows 8.1 instead of Windows version 6.3.9600. On macOS it includes the code names, on Linux it includes the distribution names and codenames if appropriate.

Usage

os_name()

Details

It uses utils::sessionInfo(), but simplifies its output a bit on Windows, to make it more concise.

Value

A character scalar.

[Package sessioninfo version 1.2.2 Index]