python_info {sessioninfo}	R Documentation

Python configuration

Description

Python configuration

Usage

python_info()

Value

Returns a reticulate::py_config object, which also has the python_info class. It is a named list of values.

If reticulate is not installed or Python is not configured, then it return a python_info object that is a character vector, and it does not have a py_config class.

Examples

python_info()
session_info(info = "all")

[Package sessioninfo version 1.2.2 Index]