sessioninfo-package {sessioninfo}	R Documentation

sessioninfo: R Session Information

Description

Query and print information about the current R session. It is similar to 'utils::sessionInfo()', but includes more information about packages, and where they were installed from.

Author(s)

Maintainer: Gábor Csárdi csardi.gabor@gmail.com

Authors:

Hadley Wickham

Winston Chang

Robert Flight

Kirill Müller

Jim Hester

Other contributors:

R Core team [contributor]

Posit Software, PBC [copyright holder, funder]

See Also

Useful links:

https://github.com/r-lib/sessioninfo#readme

https://sessioninfo.r-lib.org

Report bugs at https://github.com/r-lib/sessioninfo/issues

[Package sessioninfo version 1.2.3 Index]