Python configuration

Python configuration

python_info()

Value

Returns a reticulate::py_config object, which also has the python_info class. It is a named list of values.

If reticulate is not installed or Python is not configured, then it return a python_info object that is a character vector, and it does not have a py_config class.

Examples

python_info()
#>  Python is not available
session_info(info = "all")
#> ─ Session info  ──────────────────────────────────────────────────────────────
#>  hash: heart exclamation, person pouting: medium-light skin tone, person in tuxedo: medium-light skin tone
#> 
#>  setting  value
#>  version  R version 4.1.1 (2021-08-10)
#>  os       Ubuntu 20.04.3 LTS
#>  system   x86_64, linux-gnu
#>  ui       X11
#>  language (EN)
#>  collate  C.UTF-8
#>  ctype    C.UTF-8
#>  tz       UTC
#>  date     2021-10-31
#>  pandoc   2.7.3 @ /usr/bin/ (via rmarkdown)
#> 
#> ─ Packages ───────────────────────────────────────────────────────────────────
#>  package     * version date (UTC) lib source
#>  cachem        1.0.6   2021-08-19 [1] Custom
#>  callr         3.7.0   2021-04-20 [1] Custom
#>  cli           3.1.0   2021-10-27 [1] Custom
#>  crayon        1.4.2   2021-10-29 [1] CRAN (R 4.1.1)
#>  curl          4.3.2   2021-06-23 [1] Custom
#>  desc          1.4.0   2021-09-28 [1] Custom
#>  digest        0.6.28  2021-09-23 [1] Custom
#>  downlit       0.4.0   2021-10-29 [1] CRAN (R 4.1.1)
#>  ellipsis      0.3.2   2021-04-29 [1] Custom
#>  evaluate      0.14    2019-05-28 [1] Custom
#>  fansi         0.5.0   2021-05-25 [1] Custom
#>  fastmap       1.1.0   2021-01-25 [1] Custom
#>  fs            1.5.0   2020-07-31 [1] Custom
#>  glue          1.4.2   2020-08-27 [1] Custom
#>  htmltools     0.5.2   2021-08-25 [1] Custom
#>  httr          1.4.2   2020-07-20 [1] Custom
#>  knitr         1.36    2021-09-29 [1] Custom
#>  lifecycle     1.0.1   2021-09-24 [1] Custom
#>  magrittr      2.0.1   2020-11-17 [1] Custom
#>  memoise       2.0.0   2021-01-26 [1] Custom
#>  pillar        1.6.4   2021-10-18 [1] Custom
#>  pkgconfig     2.0.3   2019-09-22 [1] Custom
#>  pkgdown       1.6.1   2020-09-12 [1] Custom
#>  processx      3.5.2   2021-04-30 [1] Custom
#>  ps            1.6.0   2021-02-28 [1] Custom
#>  purrr         0.3.4   2020-04-17 [1] Custom
#>  R6            2.5.1   2021-08-19 [1] Custom
#>  ragg          1.2.0   2021-10-30 [1] CRAN (R 4.1.1)
#>  rematch2      2.1.2   2020-05-01 [1] Custom
#>  rlang         0.4.12  2021-10-18 [1] Custom
#>  rmarkdown     2.11    2021-09-14 [1] Custom
#>  rprojroot     2.0.2   2020-11-15 [1] Custom
#>  rstudioapi    0.13    2020-11-12 [1] Custom
#>  sessioninfo * 1.2.0   2021-10-31 [1] CRAN (R 4.1.1)
#>  systemfonts   1.0.3   2021-10-13 [1] Custom
#>  textshaping   0.3.6   2021-10-13 [1] Custom
#>  tibble        3.1.5   2021-09-30 [1] Custom
#>  utf8          1.2.2   2021-07-24 [1] Custom
#>  vctrs         0.3.8   2021-04-29 [1] Custom
#>  whisker       0.4     2019-08-28 [1] Custom
#>  withr         2.4.2   2021-04-18 [1] Custom
#>  xfun          0.27    2021-10-18 [1] Custom
#>  xml2          1.3.2   2020-04-23 [1] Custom
#>  yaml          2.2.1   2020-02-01 [1] Custom
#> 
#>  [1] /home/runner/work/_temp/Library
#>  [2] /opt/R/4.1.1/lib/R/library
#> 
#> ─ External software ──────────────────────────────────────────────────────────
#>  setting        value
#>  cairo          1.16.0
#>  cairoFT
#>  pango          1.44.7
#>  png            1.6.37
#>  jpeg           8.0
#>  tiff           LIBTIFF, Version 4.1.0
#>  tcl            8.6.10
#>  curl           7.68.0
#>  zlib           1.2.11
#>  bzlib          1.0.8, 13-Jul-2019
#>  xz             5.2.4
#>  PCRE           10.34 2019-11-21
#>  ICU            66.1
#>  TRE            TRE 0.8.0 R_fixes (BSD)
#>  iconv          glibc 2.31
#>  readline       8.0
#>  BLAS           /usr/lib/x86_64-linux-gnu/blas/libblas.so.3.9.0
#>  lapack         /usr/lib/x86_64-linux-gnu/lapack/liblapack.so.3.9.0
#>  lapack_version 3.9.0
#> 
#> ─ Python configuration ───────────────────────────────────────────────────────
#>  Python is not available
#> 
#> ──────────────────────────────────────────────────────────────────────────────