session_diff {sessioninfo}	R Documentation

Compare session information from two sources

Description

Compare session information from two sources

Usage

session_diff(old = "local", new = "clipboard", ...)

Arguments

old, new
A session_info object (the return value of session_info()), or a pointer to session_info() output. See details below.

...
Passed to any new session_info() calls.

Details

Various way to specify old and new:

A session_info object.

"local" runs session_info() in the current session, and uses its output.

"clipboard" takes the session info from the system clipboard. If the clipboard contains a URL, it is followed to download the session info.

A URL starting with http:// or https://. session_diff searches the HTML (or text) page for the session info header to find the session info.

Examples

session_diff()

[Package sessioninfo version 1.2.2 Index]