Print session information

This is utils::sessionInfo() re-written from scratch to both exclude data that's rarely useful (e.g., the full collate string or base packages loaded) and include stuff you'd like to know (e.g., where a package was installed from).

session_info(
  pkgs = c("!loaded", "!attached", "!installed")[1],
  include_base = FALSE,
  info = c("auto", "all", "platform", "packages", "python", "external"),
  dependencies = NA,
  to_file = FALSE
)

Arguments

pkgs
Which packages to show. It may be:

NULL or "!loaded": show all loaded packages,

"!attached": show all attached packages,

"!installed": show all installed packages,

a character vector of package names. Their (hard) dependencies are also shown by default, see the dependencies argument.

include_base
Include base packages in summary? By default this is false since base packages should always match the R version.

info
What information to show, it can be "auto" to choose automatically, "all" to show everything, or a character vector with elements from:

"platform": show platform information via platform_info(),

"packages": show package information via package_info(),

"python": show Python configuration via python_info(),

"external": show information about external software, via external_info().

dependencies
Whether to include the (recursive) dependencies as well. See the dependencies argument of utils::install.packages().

to_file
Whether to print the session information to a file. If TRUE the name of the file will be session-info.txt, but to_file may also be a string to specify the file name.

Value

A session_info object.

If to_file is not FALSE then it is returned invisibly. (To print it to both a file and to the screen, use (session_info(to_file = TRUE)).)

Details

Columns in the printed package list:

package: package name

*: whether the package is attached to the search path

version: package version. If the version is marked with (!) that means that the loaded and the on-disk version of the package are different.

date: when the package was built, if this information is available. This is the Date/Publication or the Built field from DESCRIPTION. (These are usually added automatically by R.) Sometimes this data is not available, then it is NA.

source: where the package was built or installed from, if available. Examples: CRAN (R 3.3.2), Github (r-lib/pkgbuild@8aab60b), Bioconductor, local.

See package_info() for the list of columns in the data frame that is returned (as opposed to printed).

Examples

session_info()
#> ─ Session info  ──────────────────────────────────────────────────────────────
#>  hash: penguin, man mechanic: light skin tone, flag: Ascension Island
#> 
#>  setting  value
#>  version  R version 4.1.1 (2021-08-10)
#>  os       Ubuntu 20.04.3 LTS
#>  system   x86_64, linux-gnu
#>  ui       X11
#>  language (EN)
#>  collate  C.UTF-8
#>  ctype    C.UTF-8
#>  tz       UTC
#>  date     2021-10-31
#>  pandoc   2.7.3 @ /usr/bin/ (via rmarkdown)
#> 
#> ─ Packages ───────────────────────────────────────────────────────────────────
#>  package     * version date (UTC) lib source
#>  cachem        1.0.6   2021-08-19 [1] Custom
#>  callr         3.7.0   2021-04-20 [1] Custom
#>  cli           3.1.0   2021-10-27 [1] Custom
#>  crayon        1.4.2   2021-10-29 [1] CRAN (R 4.1.1)
#>  curl          4.3.2   2021-06-23 [1] Custom
#>  desc          1.4.0   2021-09-28 [1] Custom
#>  digest        0.6.28  2021-09-23 [1] Custom
#>  downlit       0.4.0   2021-10-29 [1] CRAN (R 4.1.1)
#>  ellipsis      0.3.2   2021-04-29 [1] Custom
#>  evaluate      0.14    2019-05-28 [1] Custom
#>  fansi         0.5.0   2021-05-25 [1] Custom
#>  fastmap       1.1.0   2021-01-25 [1] Custom
#>  fs            1.5.0   2020-07-31 [1] Custom
#>  glue          1.4.2   2020-08-27 [1] Custom
#>  htmltools     0.5.2   2021-08-25 [1] Custom
#>  httr          1.4.2   2020-07-20 [1] Custom
#>  knitr         1.36    2021-09-29 [1] Custom
#>  lifecycle     1.0.1   2021-09-24 [1] Custom
#>  magrittr      2.0.1   2020-11-17 [1] Custom
#>  memoise       2.0.0   2021-01-26 [1] Custom
#>  pillar        1.6.4   2021-10-18 [1] Custom
#>  pkgconfig     2.0.3   2019-09-22 [1] Custom
#>  pkgdown       1.6.1   2020-09-12 [1] Custom
#>  processx      3.5.2   2021-04-30 [1] Custom
#>  ps            1.6.0   2021-02-28 [1] Custom
#>  purrr         0.3.4   2020-04-17 [1] Custom
#>  R6            2.5.1   2021-08-19 [1] Custom
#>  ragg          1.2.0   2021-10-30 [1] CRAN (R 4.1.1)
#>  rematch2      2.1.2   2020-05-01 [1] Custom
#>  rlang         0.4.12  2021-10-18 [1] Custom
#>  rmarkdown     2.11    2021-09-14 [1] Custom
#>  rprojroot     2.0.2   2020-11-15 [1] Custom
#>  rstudioapi    0.13    2020-11-12 [1] Custom
#>  sessioninfo * 1.2.0   2021-10-31 [1] CRAN (R 4.1.1)
#>  systemfonts   1.0.3   2021-10-13 [1] Custom
#>  textshaping   0.3.6   2021-10-13 [1] Custom
#>  tibble        3.1.5   2021-09-30 [1] Custom
#>  utf8          1.2.2   2021-07-24 [1] Custom
#>  vctrs         0.3.8   2021-04-29 [1] Custom
#>  whisker       0.4     2019-08-28 [1] Custom
#>  withr         2.4.2   2021-04-18 [1] Custom
#>  xfun          0.27    2021-10-18 [1] Custom
#>  xml2          1.3.2   2020-04-23 [1] Custom
#>  yaml          2.2.1   2020-02-01 [1] Custom
#> 
#>  [1] /home/runner/work/_temp/Library
#>  [2] /opt/R/4.1.1/lib/R/library
#> 
#> ──────────────────────────────────────────────────────────────────────────────
session_info("sessioninfo")
#> ─ Session info  ──────────────────────────────────────────────────────────────
#>  hash: keycap: 9, microbe, Gemini
#> 
#>  setting  value
#>  version  R version 4.1.1 (2021-08-10)
#>  os       Ubuntu 20.04.3 LTS
#>  system   x86_64, linux-gnu
#>  ui       X11
#>  language (EN)
#>  collate  C.UTF-8
#>  ctype    C.UTF-8
#>  tz       UTC
#>  date     2021-10-31
#>  pandoc   2.7.3 @ /usr/bin/ (via rmarkdown)
#> 
#> ─ Packages ───────────────────────────────────────────────────────────────────
#>  package     * version date (UTC) lib source
#>  cli           3.1.0   2021-10-27 [1] Custom
#>  glue          1.4.2   2020-08-27 [1] Custom
#>  sessioninfo * 1.2.0   2021-10-31 [1] CRAN (R 4.1.1)
#> 
#>  [1] /home/runner/work/_temp/Library
#>  [2] /opt/R/4.1.1/lib/R/library
#> 
#> ──────────────────────────────────────────────────────────────────────────────